Remove "Name" property from Player
The initial idea was to update/get the current user name live from the SteamID as it might change in the future.
However, currently the webui uses hardcoded "real names" and it would actually make the user experience worse if the names changed after the user presses the Scramble button.

This feature might be interesting again once we allow adding players during runtime.
This removal will be linked to the corresponding improvement issue in GitHub so that the code can be re-used if/when we implement this feature.