Initialize players statistics with game coordinator information
The players, via their SteamIDs, are added and each gets assigned their "initial side"', i.e., the team they played the first half on.
This is based on the - usually true - assumption that the first 5 players in the game coordinator match details are the initial CT side.
This information will then be used in the demo reader to check if this assumption is correct.
If not, also the assigned CT and T scores are incorrect, as the game coordinator has apparrently flipped the teams for this match.